Fix the text-gen issue caused by import transformer_engine.
